Move storybook files into their respective component folders
