+++ app/src/ui/link/stories/link.stories.tsx
@@ -1,10 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { text, boolean } from '@storybook/addon-knobs';
-import { action } from '@storybook/addon-actions';
 
-import { StoryGroup } from './stories';
-import { Link, ButtonLink, CapsuleLink } from '../ui/link/link';
+import { Link, ButtonLink, CapsuleLink } from '../link';
 
 storiesOf('UI/Link', module)
   .add('Link', () => (
