+++ app/src/ui/nav/navbar/horizontal_menu/horizontal_menu.module.css
@@ -15,7 +15,7 @@
 }
 
 .linkContainer {
-    width: 75px;
+    width: 200px;
     display: flex;
     justify-content: center;
 }
